backend/run.py
#!/usr/bin/env python3
"""
AI-Powered Communication Assistant
Main entry point for the application
"""

import sys
import os
import uvicorn
from dotenv import load_dotenv

# Add the app directory to Python path
sys.path.append(os.path.join(os.path.dirname(__file__), 'app'))

load_dotenv()

if __name__ == "__main__":
    # Development server configuration
    uvicorn.run(
        "app.main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )
